chore(chronograf): delete dead code
